Rename mapGeoLocation to LocationControl
